+++ internal/provider/provider.go
@@ -2,21 +2,13 @@ package provider
 
 import (
 	"context"
-	"fmt"
-	"os"
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-const (
-	EnvVarJQCliPath = "JQ_CLI_PATH"
 )
 
 func New() tfsdk.Provider {
 	return &provider{}
-
 }
 
 type provider struct {
@@ -36,20 +28,12 @@ structures... without messing in unreadable HCL code.
 ## Terraform cloud
 
 The provider is portable and doesn't depend on any binary, so its compatible with terraform cloud workers.`,
-		Attributes: map[string]tfsdk.Attribute{
-			"jq_cli_path": {
-				Type:        types.StringType,
-				Optional:    true,
-				Description: fmt.Sprintf("The path that points to the JQ cli binary. Also `%s` env var can be used. (by default `jq` on system path, ignored if run in Terraform cloud).", EnvVarJQCliPath),
-			},
-		},
+		Attributes: map[string]tfsdk.Attribute{},
 	}, nil
 }
 
 // Provider configuration.
-type providerData struct {
-	JQCliPath types.String `tfsdk:"jq_cli_path"`
-}
+type providerData struct{}
 
 // This is like if it was our main entrypoint.
 func (p *provider) Configure(ctx context.Context, req tfsdk.ConfigureProviderRequest, resp *tfsdk.ConfigureProviderResponse) {
@@ -61,29 +45,7 @@ func (p *provider) Configure(ctx context.Context, req tfsdk.ConfigureProviderReq
 		return
 	}
 
-	// Error summaries
-	const (
-		configErrSummary = "Unable to configure client"
-	)
-
-	_, err := p.configureJQCliPath(config)
-	if err != nil {
-		resp.Diagnostics.AddError(configErrSummary, "Invalid JQ cli path:\n\n"+err.Error())
-	}
-
 	p.configured = true
-}
-
-func (p *provider) configureJQCliPath(config providerData) (string, error) {
-	// If not set get from env, the value has priority.
-	var cliPath string
-	if config.JQCliPath.Null {
-		cliPath = os.Getenv(EnvVarJQCliPath)
-	} else {
-		cliPath = config.JQCliPath.Value
-	}
-
-	return cliPath, nil
 }
 
 func (p *provider) GetResources(_ context.Context) (map[string]tfsdk.ResourceType, diag.Diagnostics) {