+++ CHANGELOG.md
@@ -2,5 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Bootstrap provider.
+- JQ data source.
+
 [unreleased]: https://github.com/slok/terraform-provider-dataprocessor/compare/v0.1.0...HEAD
 [v0.1.0]: https://github.com/slok/terraform-provider-dataprocessor/releases/tag/v0.1.0